Change identification of AST nodes
Annotations would be compared for equality or printed before, but now
there is a separate typeclass IdentifiableAnn that determines what the
contribution of a particular annotation is.

For example, Decor in Delta annotation for predicates do contribute but
not SrcSpan in the Base annotation. Thus far it is only used for Eq and
Ord.