Distinguish Current annotation from Constant in Delta
When computing deltas, both intensional predicates whose result we are
interested in and the extensional predicates are stored together.
However, the extensional predicates when they become Prev through aging
during semi-naive evaluation their Normal is not updated (that only
happens for Normals). So we distinguish Normal into Current and
Constant for intensional results and extensionals (relative)
respectively.